pyproject.toml
[project]
name = "yifang"
version = "0.1.0"
description = "蚁坊 (Yifang) — AI 乙方公司框架"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=2.29.0",
    "python-dotenv>=1.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-mock>=3.12",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/yifang"]

[tool.pytest.ini_options]
testpaths = ["tests"]
